+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Do not execute YAML option lines (enables execution of cell magics)
 - Don't require a cell selection for 'Run All Cells'
+- Run Python cells as distinct commands in Interactive Console
 
 ## 1.33.0 (Release on 24 August 2022)